Build 1.5.5
Changing SarCalculation sensitivity calculation to default on a linear fit. If the R2 is below 0.98, then the calculator will attempt the curve fit.

Added a calculateSensitivityLine API to give sensitivity values for a linear fit given a testdata _id. This was not created to use in production, but for engineers.